requirements.txt
ipython>=8.10
jupyterlab>=3.0
notebook
kedro~=1.0.0
kedro-datasets[pandas]>=4.1.0
pandas>=2.3.2
numpy>=2.0.2
scikit-learn>=1.6.1
ipykernel>=6.30.1
matplotlib>=3.9.4
seaborn>=0.13.2
jupyter>=1.1.1
imbalanced-learn>=0.11.0
scipy>=1.11.0
xgboost>=2.0.0
dvc>=3.0.0
umap-learn>=0.5.0
plotly>=5.0.0
hdbscan>=0.8.0
